added Pvalue corrected for NDoF in BCModel; small update in BCMath

+++ BAT/BCMath.h
@@ -25,7 +25,7 @@
 #define BCMATH_NFACT_ALIMIT 20
 
 // ---------------------------------------------------------
-#include <cstring>
+//#include <cstring>
 
 namespace BCMath
 {
@@ -125,7 +125,7 @@ namespace BCMath
 	 * Calculates the logarithm of chi square function:
 	 * chi2(double x; size_t n)
 	 */
-	double LogChi2(double x, size_t n);
+	double LogChi2(double x, int n);
 
 	/**
 	 * Calculates the logarithm of normalized voigtian function:
@@ -135,8 +135,8 @@ namespace BCMath
 	 *  gaussian(x) = 1/(sqrt(2*pi)*sigma) * exp(x*x/(2*sigma*sigma)
 	 *    and
 	 *  lorentz(x) = (1/pi)*(gamma/2) / (x*x + (gamma/2)*(gamma/2))
-	 * 
-	 * it is singly peaked at x=0. 
+	 *
+	 * it is singly peaked at x=0.
 	 * The width of the peak is decided by sigma and gamma,
 	 * so they should be positive.
 	 */
